pyproject.toml
[project]
name = "realtime-counselor-agent"
version = "0.1.0"
description = "Real-time counseling agent with WebRTC audio streaming and Google STT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-speech>=2.30.0",
    "websockets>=14.1",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "aiortc>=1.14.0",
    "aiohttp>=3.9.0",
    "numpy>=2.3.5",
    "httpx>=0.25.0",
    "langchain[openai]>=0.1.0",
    "langgraph>=0.0.20",
    "langgraph-cli[inmem]>=0.4.7",
    "langsmith>=0.4.42",
    "openai[aiohttp]>=2.8.1",
    "python-multipart>=0.0.6",
    "asyncpg>=0.31.0",
    "redis>=7.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "langchain-openai>=0.0.5",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "langchain-huggingface>=0.0.1",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "typing-extensions>=4.9.0",
    "langchain-postgres>=0.0.16",
    "psycopg>=3.3.2",
    "beautifulsoup4>=4.14.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
